+++ inst/doc/wrapr_applicable.R
@@ -14,10 +14,24 @@ library("wrapr")
 
 ## ----sinfn---------------------------------------------------------------
 function_reference <- list(f = sin)
-class(function_reference) <- "wrapr_applicable"
-function_reference$wrapr_function <- function(pipe_left_arg, 
-                                              pipe_right_arg, 
-                                              pipe_environment) {
+class(function_reference) <- c("wrapr_applicable", "ourclass")
+
+#' Wrapr function.
+#'
+#' S3 dispatch on tyhpe of pipe_right_argument.
+#'
+#' @param pipe_left_arg left argument.
+#' @param pipe_right_arg right argument.
+#' @param pipe_environment environment to evaluate in.
+#' @param pipe_name character, name of pipe operator.
+#' @return result
+#'
+#' @export
+#'
+wrapr_function.ourclass <- function(pipe_left_arg, 
+                                    pipe_right_arg,
+                                    pipe_environment,
+                                    pipe_name) {
   pipe_right_arg$f(pipe_left_arg)
 }
 
@@ -26,5 +40,23 @@ function_reference
 5 %.>% function_reference
 
 function_reference$f <- sqrt
+5 %.>% function_reference
+
+## ----debug---------------------------------------------------------------
+wrapr_function.ourclass <- function(pipe_left_arg, 
+                                    pipe_right_arg,
+                                    pipe_environment,
+                                    pipe_name) {
+  print("pipe_left_arg")
+  print(pipe_left_arg)
+  print("pipe_right_arg")
+  print(pipe_right_arg)
+  print("pipe_environment")
+  print(pipe_environment)
+  print("pipe_name")
+  print(pipe_name)
+  pipe_right_arg$f(pipe_left_arg)
+}
+
 5 %.>% function_reference